+++ app/src/main/java/com/example/facturas/data/network/invoicesApi/models/InvoiceResponse.kt
@@ -3,7 +3,7 @@ package com.example.facturas.data.network.invoicesApi.models
 data class InvoiceResponse(
     val descEstado: String,
     val fecha: String,
-    val importeOrdenacion: Double
+    val importeOrdenacion: Float
 ) {
     fun asApiModel(): InvoiceApiModel {
         return InvoiceApiModel(
